+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 
 * Signal an error sooner if the user misconfigured `cider-known-endpoints`.
 * `cider-inspect-read-and-inspect` is obsolete. Use `cider-inspect-expression` instead.
+* Extremely long overlays are truncated and `cider-inspect-last-result` is recommended.
 
 ### Bugs fixed